Display all Twitter pictures, hide all user embeds
Behavior had to change because now Twitter embeds can contain all pictures in one single embed.
This change makes it so that the original message has its embed nuked, and now the first entry in media
gets posted along the other 3, for a total of 4 (the maximum a tweet can have).

Additionally, the footer now states it's a viewer to remind users it's not meant to be a nuissance.