Rework roll syntax
The !roll command syntax has been completely revamped. It no longer requires point additions to be made after a roll. It is now more strict about its usage to help prevent it from silently failing or miscalculating. Each roll or point
addition has to be explicitly stated, and the bot will always reply with
a short breakdown of the logic it went through to arrive to its result.

The maximum number of dice per pip roll has been lowered to 100 down from 10000.

Rolls can now be given a sign and substract from the total.