Pixiv embeds: Now limited to thumbnail preview
It's proven unfeasible to keep posting whole galleries when users can
just spam links and drive everyone crazy with notifications, as such the
bot is strictly limited to sending only one picture as opposed to many
as before. This feature could be turned back on by using the bang syntax
that was introduced with channel linking, but it would still bring back
the same problem as before.

To make up for this change, the bot will now override the preview sent,
nuking the original message's while trying to pass as the real one,
albeit improved.