Galleries: Format numbers;  Even out embed style
Gallery embeds all now should have proper number formatting, that is to
say, they are comma-separated where appropiate. Additionally Pixiv
embeds now feature favorites and view counts.

There was an attempt to add in avatar embeds, but it's impossible to
achieve as discord embeds require an url. The avatars would need to be
uploaded to a different website but that would probably be too involved
to achieve.