Emded nuke: Handle missing permissions
When a user posts an url with an upgradable embed, and the bot doesn't
have permissions to hide the redundant preview from the url that
triggered it, then it will silently continue. The fix can be solved
using a common library (cog, etc.) but for now the current
implementation gets the job done.