Deviantart embeds fixes: Suppression, desc bug
Move message embed suppression code after the part where the embed has been retrieved. This is due to the fact that the embed may fail and in such cases the existing embed will be preferred. Also fixed a bug related to descriptions being missing in posts that caused embedss to not be sent and thus causing message suppression without any embeds as mentioned above.