DeviantArt previews: Add "smart" gallery detection
Sending multiple links from DeviantArt in a single message now checks
whether or not the links are "similar" from each other. If they are,
then the bot will proceed to "merge" the embeds together into one batch,
for a maximum of 5.

The smart gallery detection doesn't account for the similarity between
all urls, only the average of the comparison of the first to every other
consecutive link.