Implement EmbedGroup class
This class attempts to simplify the management of multiple embeds.
Currently they only pass over properties set to the first and last
embeds within an embed group, and is capable of applying the same
properties to other embeds it contains.

It's almost a full drop-in replacement that doesn't change how it shows
up to the user (for now).