Rename previews to sites
Encapsulating previews to be able to interface with the sites and only
being able to do one thing doesn't feel like a good idea to me. Instead
of restricting their use to only previews, they're now sites which can
do any operation they're capable if called for. This is a breaking
change.