Previews: Use tweepy async and twitter API v2
Now Twitter requests shouldn't be blocking, although its effects were
never really felt.

The upgrade to the v2 API caused us to be unable to
figure out where a tweet's text starts and ends reliably. We have a
workaround going, but it's not a really good approach (or so it seems).