Add DALL-E mini/DALL-E mega support
This is an attempt to directly implement dalle. Currently there's no way
to get this running because jax is throwing a random error and we have
no clue why this is happening, but it's a great start. However, more
research should be done to get this up and running (and a powerful
machine as well).