Music: Play video sent if a playlist was included
There was a bug that made it so that the video chosen to be played was
always the first element from the playlist regardless of the video url
sent to the bot. Now it should play the video sent although it's slow
because currently the entire playlist is downloaded first.