DA: Fix missing csrf_token error
DA now asks for a csrf token to be passed along with requests, as well
as cookies. This small patch fixes it but there are some areas that can
be optimized, like caching the cookies and the csrf token itself, since
as of now it fetches it every new instance.