tests, !roll: Set up not DRY unit testing
Finally!!! We have some testing going on now. Sadly we haven't figured out how to do it properly. Because we use discord.py, we can't freely have the tests import the cogs since they require bot data.

Additionally, the logic of the !roll command has been separated from the command itself. It still lives in the cog because we want the code to be reloadable while the bot is already running.