!roll: Minor speed improvements, better tests
Apparently string concatenation is a rather expensive operation. Therefore we're using the recommended way of doing it: with lists. Other parts of the bot may follow suit.

As for the test, we've added more checks to make sure it works fine but it's frankly way too much text. Perhaps we should focus hard on checking whether or not the roll parser undertood the command correctly, and check for the message logs second. In any case, the message logs should be done with mocked rolls because finding lucky seeds is not a very reasonable thing to do.